move preview setting listener over to `createConfigChangeListener()`
